GameController
Added GameController script that keeps track of amount of enemies. Once it hits 0, text appears that says "Mission Success" and you are dropped in a new scene that will be used for an upgrade screen. You can then move in any direction to be dropped back in the Game scene